separate buttons out to separate sheet along with quality functions
