+++ lib/createToDo.js
@@ -68,7 +68,7 @@ module.exports = function createIdeaBox(ideabox){
 // module.exports = ideaViewHelper;
 //
 // //index.js
-
+//
 // const ideaViewHelper = require...
-
+//
 // ideaViewHelper.createIdeaBox()
